+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 # main
 
+Save optimized thresholds as yaml instead of pickle to make it easier to read
+
 ### 1.10.2
 - Add support for metadata propagation through compound pdal pipelines:
   - fix epsg propagation